+++ .devcontainer/test.sh
@@ -32,13 +32,11 @@ answers:
   identity.name: Test User
   identity.email: test@example.com
   hat.1.name: personal
-  hat.1.aws_profile: default
   hat.1.kube_context: ""
   hat.add.2: true
   hat.2.name: client
   hat.2.git_name: Test Client
   hat.2.git_email: test@client.example
-  hat.2.aws_profile: client-account
   hat.2.kube_context: client
   hat.add.3: false
   secrets.enabled: false